version commentée qui marche pour mael
