+++ src/main/java/com/example/geodevineur/GeoDevineurApplication.java
@@ -5,15 +5,15 @@ import java.util.Random;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import com.example.geodevineur.condition.CardinalCond;
-import com.example.geodevineur.condition.Condition;
-import com.example.geodevineur.condition.ContainLetterCond;
-import com.example.geodevineur.condition.NeighbourCond;
-import com.example.geodevineur.condition.PoliticCond;
-import com.example.geodevineur.condition.SeasideCond;
-import com.example.geodevineur.tables.Departement;
-import com.example.geodevineur.enumerations.Cardinal;
-import com.example.geodevineur.enumerations.Politic;
+//import com.example.geodevineur.condition.CardinalCond;
+//import com.example.geodevineur.condition.Condition;
+//import com.example.geodevineur.condition.ContainLetterCond;
+//import com.example.geodevineur.condition.NeighbourCond;
+//import com.example.geodevineur.condition.PoliticCond;
+//import com.example.geodevineur.condition.SeasideCond;
+//import com.example.geodevineur.tables.Departement;
+//import com.example.geodevineur.enumerations.Cardinal;
+//import com.example.geodevineur.enumerations.Politic;
 
 @SpringBootApplication
 public class GeoDevineurApplication {
@@ -24,21 +24,21 @@ public class GeoDevineurApplication {
     }
 
     /* Génération aléatoire de la question (restreignant le champs des possibles) */
-    public Condition<Departement> questionChoice(Departement solution, int possibilites) {
-        /* On choisit aléatoirement le type de question */
+    /*public Condition<Departement> questionChoice(Departement solution, int possibilites) {
+        *//* On choisit aléatoirement le type de question *//*
         Random random = new Random();
         int randomInt = random.nextInt(9);
 
         Condition<Departement> cond;
 
         switch (randomInt) {
-            /* Puis on définit l'argument de sorte à ce que la question restreigne le champs des possibles
+            *//* Puis on définit l'argument de sorte à ce que la question restreigne le champs des possibles
              * Si ce n'est pas possible, alors on choisit une autre question
-             */
-            /* Cardinal */
+             *//*
+            *//* Cardinal *//*
             case 0:
                 Cardinal c = solution.getRegion().getCardinal();
-                /* compter le nb de département qui ont ce cardinal */
+                *//* compter le nb de département qui ont ce cardinal *//*
                 if (nbMatch < possibilites) {
                     return new CardinalCond<Departement>(solution);
                 }
@@ -49,16 +49,16 @@ public class GeoDevineurApplication {
 
                 break;
 
-            /* Contient telle lettre */
+            *//* Contient telle lettre *//*
             case 1:
-                /* On prend un caractère du nom au hasard
+                *//* On prend un caractère du nom au hasard
                  * AU LIEU DE PRENDRE AU HASARD IL FAUDRAIT EN PRENDRE UN QUI RESTREINT
                  * PASSER EN PARAM DES CONDITIONS LE NB DE POSSIBILITES ?
-                 */
+                 *//*
                 ContainLetterCond<Departement> letterCond = new ContainLetterCond<Departement>(solution);
                 char letter = letterCond.getLetter();
 
-                /* compter le nb de département qui ont cette lettre dans la BDD */
+                *//* compter le nb de département qui ont cette lettre dans la BDD *//*
                 if (nbMatch < possibilites) {
                     return letterCond;
                 }
@@ -69,10 +69,10 @@ public class GeoDevineurApplication {
 
                 break;
 
-            /* Num du département */
-            /* PAS CAPTE LE SYSTEME DE CHOIX DE MAEL */
+            *//* Num du département *//*
+            *//* PAS CAPTE LE SYSTEME DE CHOIX DE MAEL *//*
             case 2:
-                /* compter le nb de département qui sont inférieurs/supérieurs à ce num */
+                *//* compter le nb de département qui sont inférieurs/supérieurs à ce num *//*
                 if (nbMatch < possibilites) {
                     return new CardinalCond<Departement>(solution);
                 }
@@ -83,10 +83,10 @@ public class GeoDevineurApplication {
 
                 break;
 
-            /* Nb char */
-            /* VOIR COMMENT DEFINIR LES PALLIERS */
+            *//* Nb char *//*
+            *//* VOIR COMMENT DEFINIR LES PALLIERS *//*
             case 3:
-                /* compter le nb de département qui ont plus ou moins que X lettres */
+                *//* compter le nb de département qui ont plus ou moins que X lettres *//*
                 if (nbMatch < possibilites) {
                     return new CardinalCond<Departement>(solution);
                 }
@@ -97,11 +97,11 @@ public class GeoDevineurApplication {
 
                 break;
 
-            /* Nb voisins */
+            *//* Nb voisins *//*
             case 4:
-                /* PAS CAPTE LE SYSTEME DE CHOIX DE MAEL */
+                *//* PAS CAPTE LE SYSTEME DE CHOIX DE MAEL *//*
                 int voisins = solution.getNeightbours();
-                /* compter le nb de département qui ont ce nb de voisins */
+                *//* compter le nb de département qui ont ce nb de voisins *//*
                 if (nbMatch < possibilites) {
                     return new NeighbourCond<Departement>(solution);
                 }
@@ -112,10 +112,10 @@ public class GeoDevineurApplication {
 
                 break;
 
-            /* Politique */
+            *//* Politique *//*
             case 5:
                 Politic p = solution.getPolitic();
-                /* compter le nb de département qui ont cette politique */
+                *//* compter le nb de département qui ont cette politique *//*
                 if (nbMatch < possibilites) {
                     return new PoliticCond<Departement>(solution);
                 }
@@ -126,10 +126,10 @@ public class GeoDevineurApplication {
 
                 break;
 
-            /* Population */
-            /* PAS CAPTE LE SYSTEME DE CHOIX DE MAEL */
+            *//* Population *//*
+            *//* PAS CAPTE LE SYSTEME DE CHOIX DE MAEL *//*
             case 6:
-                /* compter le nb de département qui match le palier */
+                *//* compter le nb de département qui match le palier *//*
                 if (nbMatch < possibilites) {
                     return new CardinalCond<Departement>(solution);
                 }
@@ -140,10 +140,10 @@ public class GeoDevineurApplication {
 
                 break;
 
-            /* Cotier */
+            *//* Cotier *//*
             case 7:
                 boolean cote = solution.isSeaside();
-                /* compter le nb de département qui sont pareils */
+                *//* compter le nb de département qui sont pareils *//*
                 if (nbMatch < possibilites) {
                     return new SeasideCond<Departement>(solution);
                 }
@@ -154,10 +154,10 @@ public class GeoDevineurApplication {
 
                 break;
 
-            /* Surface */
-            /* PAS CAPTE LE SYSTEME DE CHOIX DE MAEL */
+            *//* Surface *//*
+            *//* PAS CAPTE LE SYSTEME DE CHOIX DE MAEL *//*
             case 8:
-                /* compter le nb de département qui match ce palier */
+                *//* compter le nb de département qui match ce palier *//*
                 if (nbMatch < possibilites) {
                     return new CardinalCond<Departement>(solution);
                 }
@@ -172,6 +172,6 @@ public class GeoDevineurApplication {
                 break;
         }
 
-    }
+    }*/
 
 }
